Struct google_cloudbilling1::BillingAccountMethods

pub struct BillingAccountMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on billingAccount resources. It is not used directly, but through the Cloudbilling hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_cloudbilling1 as cloudbilling1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use cloudbilling1::Cloudbilling;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Cloudbilling::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create(...)`, `get(...)`, `get_iam_policy(...)`, `list(...)`, `patch(...)`, `projects_list(...)`, `set_iam_policy(...)` and `test_iam_permissions(...)`
// to build up your call.
let rb = hub.billing_accounts();

Methods

impl<'a, C, A> BillingAccountMethods<'a, C, A>

pub fn create(
&self,
request: BillingAccount
) -> BillingAccountCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a billing account. This method can only be used to create billing subaccounts by GCP resellers. When creating a subaccount, the current authenticated user must have the billing.accounts.update IAM permission on the master account, which is typically given to billing account administrators. This method will return an error if the master account has not been provisioned as a reseller account.

Arguments

request - No description provided.

pub fn list(&self) -> BillingAccountListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the billing accounts that the current authenticated user has permission to view.

pub fn get(&self, name: &str) -> BillingAccountGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets information about a billing account. The current authenticated user must be a viewer of the billing account.

Arguments

name - The resource name of the billing account to retrieve. For example, billingAccounts/012345-567890-ABCDEF.

pub fn set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> BillingAccountSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the access control policy for a billing account. Replaces any existing policy. The caller must have the billing.accounts.setIamPolicy permission on the account, which is often given to billing account administrators.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn patch(
&self,
request: BillingAccount,
name: &str
) -> BillingAccountPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a billing account's fields. Currently the only field that can be edited is display_name. The current authenticated user must have the billing.accounts.update IAM permission, which is typically given to the administrator of the billing account.

Arguments

request - No description provided.

name - The name of the billing account resource to be updated.

pub fn projects_list(
&self,
name: &str
) -> BillingAccountProjectListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the projects associated with a billing account. The current authenticated user must have the billing.resourceAssociations.list IAM permission, which is often given to billing account viewers.

Arguments

name - The resource name of the billing account associated with the projects that you want to list. For example, billingAccounts/012345-567890-ABCDEF.

pub fn test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> BillingAccountTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Tests the access control policy for a billing account. This method takes the resource and a set of permissions as input and returns the subset of the input permissions that the caller is allowed for that resource.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn get_iam_policy(
&self,
resource: &str
) -> BillingAccountGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the access control policy for a billing account. The caller must have the billing.accounts.getIamPolicy permission on the account, which is often given to billing account viewers.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.