Struct google_cloudbilling1::ProjectBillingInfo

pub struct ProjectBillingInfo {
    pub project_id: Option<String>,
    pub name: Option<String>,
    pub billing_enabled: Option<bool>,
    pub billing_account_name: Option<String>,
}

Encapsulation of billing information for a Developers Console project. A project has at most one associated billing account at a time (but a billing account can be assigned to multiple projects).

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update billing info projects (request|response)

get billing info projects (response)

Fields
project_id: Option<String>

The ID of the project that this ProjectBillingInfo represents, such as tokyo-rain-123. This is a convenience field so that you don't need to parse the name field to obtain a project ID. This field is read-only.

name: Option<String>

The resource name for the ProjectBillingInfo; has the form projects/{project_id}/billingInfo. For example, the resource name for the billing information for project tokyo-rain-123 would be projects/tokyo-rain-123/billingInfo. This field is read-only.

billing_enabled: Option<bool>

True if the project is associated with an open billing account, to which usage on the project is charged. False if the project is associated with a closed billing account, or no billing account at all, and therefore cannot use paid services. This field is read-only.

billing_account_name: Option<String>

The resource name of the billing account associated with the project, if any. For example, billingAccounts/012345-567890-ABCDEF.